+++ Unidad2/Practicas/Gradient_Boosted_Tree _Classifier/Gradient-boosted_tree_classifier.scala
@@ -1,36 +1,36 @@
-// Equipo #2
+// Equipo #4       Gradient-boosted tree classifier
 
-// Decision tree classifier
 import org.apache.spark.ml.Pipeline
-import org.apache.spark.ml.classification.DecisionTreeClassificationModel
-import org.apache.spark.ml.classification.DecisionTreeClassifier
+import org.apache.spark.ml.classification.{GBTClassificationModel, GBTClassifier}
 import org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator
 import org.apache.spark.ml.feature.{IndexToString, StringIndexer, VectorIndexer}
 // $example off$
 import org.apache.spark.sql.SparkSession
 
-val spark = SparkSession.builder.appName("DecisionTreeClassificationExample").getOrCreate()
+val spark = SparkSession.builder.appName("GradientBoostedTreeClassifierExample").getOrCreate()
+
     // $example on$
-    // Load the data stored in LIBSVM format as a DataFrame.
+    // Load and parse the data file, converting it to a DataFrame.
     val data = spark.read.format("libsvm").load("/home/armando/Escritorio/spark-2.4.4-bin-hadoop2.7/data/mllib/sample_libsvm_data.txt")
 
     // Index labels, adding metadata to the label column.
     // Fit on whole dataset to include all labels in index.
     val labelIndexer = new StringIndexer().setInputCol("label").setOutputCol("indexedLabel").fit(data)
     // Automatically identify categorical features, and index them.
+    // Set maxCategories so features with > 4 distinct values are treated as continuous.
     val featureIndexer = new VectorIndexer().setInputCol("features").setOutputCol("indexedFeatures").setMaxCategories(4).fit(data)
-    featureIndexer.show()
+
     // Split the data into training and test sets (30% held out for testing).
     val Array(trainingData, testData) = data.randomSplit(Array(0.7, 0.3))
 
-    // Train a DecisionTree model.
-    val dt = new DecisionTreeClassifier().setLabelCol("indexedLabel").setFeaturesCol("indexedFeatures")
+    // Train a GBT model.
+    val gbt = new GBTClassifier().setLabelCol("indexedLabel").setFeaturesCol("indexedFeatures").setMaxIter(10).setFeatureSubsetStrategy("auto")
 
     // Convert indexed labels back to original labels.
     val labelConverter = new IndexToString().setInputCol("prediction").setOutputCol("predictedLabel").setLabels(labelIndexer.labels)
 
-    // Chain indexers and tree in a Pipeline.
-    val pipeline = new Pipeline().setStages(Array(labelIndexer, featureIndexer, dt, labelConverter))
+    // Chain indexers and GBT in a Pipeline.
+    val pipeline = new Pipeline().setStages(Array(labelIndexer, featureIndexer, gbt, labelConverter))
 
     // Train model. This also runs the indexers.
     val model = pipeline.fit(trainingData)
@@ -44,8 +44,7 @@ val spark = SparkSession.builder.appName("DecisionTreeClassificationExample").ge
     // Select (prediction, true label) and compute test error.
     val evaluator = new MulticlassClassificationEvaluator().setLabelCol("indexedLabel").setPredictionCol("prediction").setMetricName("accuracy")
     val accuracy = evaluator.evaluate(predictions)
-    println(s"Test Error = ${(1.0 - accuracy)}")
+    println(s"Test Error = ${1.0 - accuracy}")
 
-    val treeModel = model.stages(2).asInstanceOf[DecisionTreeClassificationModel]
-    println(s"Learned classification tree model:\n ${treeModel.toDebugString}")
-    // $example off$
+    val gbtModel = model.stages(2).asInstanceOf[GBTClassificationModel]
+    println(s"Learned classification GBT model:\n ${gbtModel.toDebugString}")
