+++ Unidad2/Practicas/Random_Forest_Classifier/RandomForestClassifier.scala
@@ -1,36 +1,36 @@
-// Equipo #2
+// Equipo # 3 Random forest classifier
 
-// Decision tree classifier
 import org.apache.spark.ml.Pipeline
-import org.apache.spark.ml.classification.DecisionTreeClassificationModel
-import org.apache.spark.ml.classification.DecisionTreeClassifier
+import org.apache.spark.ml.classification.{RandomForestClassificationModel, RandomForestClassifier}
 import org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator
 import org.apache.spark.ml.feature.{IndexToString, StringIndexer, VectorIndexer}
 // $example off$
 import org.apache.spark.sql.SparkSession
 
-val spark = SparkSession.builder.appName("DecisionTreeClassificationExample").getOrCreate()
+val spark = SparkSession.builder.appName("RandomForestClassifierExample").getOrCreate()
+
     // $example on$
-    // Load the data stored in LIBSVM format as a DataFrame.
+    // Load and parse the data file, converting it to a DataFrame.
     val data = spark.read.format("libsvm").load("/home/armando/Escritorio/spark-2.4.4-bin-hadoop2.7/data/mllib/sample_libsvm_data.txt")
 
     // Index labels, adding metadata to the label column.
     // Fit on whole dataset to include all labels in index.
     val labelIndexer = new StringIndexer().setInputCol("label").setOutputCol("indexedLabel").fit(data)
     // Automatically identify categorical features, and index them.
+    // Set maxCategories so features with > 4 distinct values are treated as continuous.
     val featureIndexer = new VectorIndexer().setInputCol("features").setOutputCol("indexedFeatures").setMaxCategories(4).fit(data)
-    featureIndexer.show()
+
     // Split the data into training and test sets (30% held out for testing).
     val Array(trainingData, testData) = data.randomSplit(Array(0.7, 0.3))
 
-    // Train a DecisionTree model.
-    val dt = new DecisionTreeClassifier().setLabelCol("indexedLabel").setFeaturesCol("indexedFeatures")
+    // Train a RandomForest model.
+    val rf = new RandomForestClassifier().setLabelCol("indexedLabel").setFeaturesCol("indexedFeatures").setNumTrees(10)
 
     // Convert indexed labels back to original labels.
     val labelConverter = new IndexToString().setInputCol("prediction").setOutputCol("predictedLabel").setLabels(labelIndexer.labels)
 
-    // Chain indexers and tree in a Pipeline.
-    val pipeline = new Pipeline().setStages(Array(labelIndexer, featureIndexer, dt, labelConverter))
+    // Chain indexers and forest in a Pipeline.
+    val pipeline = new Pipeline().setStages(Array(labelIndexer, featureIndexer, rf, labelConverter))
 
     // Train model. This also runs the indexers.
     val model = pipeline.fit(trainingData)
@@ -46,6 +46,5 @@ val spark = SparkSession.builder.appName("DecisionTreeClassificationExample").ge
     val accuracy = evaluator.evaluate(predictions)
     println(s"Test Error = ${(1.0 - accuracy)}")
 
-    val treeModel = model.stages(2).asInstanceOf[DecisionTreeClassificationModel]
-    println(s"Learned classification tree model:\n ${treeModel.toDebugString}")
-    // $example off$
+    val rfModel = model.stages(2).asInstanceOf[RandomForestClassificationModel]
+    println(s"Learned classification forest model:\n ${rfModel.toDebugString}")
